move deps into control_flow
