+++ jsx/control_flow/show.tsx
@@ -1,9 +1,11 @@
-import { html, HtmlEscapedString } from "../deps.ts";
+import { html, HtmlEscapedString } from "./deps.ts";
 
 /** [TODO](https://www.solidjs.com/docs/latest/api#show) */
-export function Show<T>(
-    { when, fallback, children }: ShowProps<T>,
-): HtmlEscapedString {
+export function Show<T>({
+    when,
+    fallback,
+    children,
+}: ShowProps<T>): HtmlEscapedString {
     if (!when) return html`${fallback}`;
     return html`${typeof children === "function" ? children(when) : children}`;
 }
