--- jsx/deps.ts
@@ -1,4 +1,0 @@
-// hono
-export { JSXNode } from "https://deno.land/x/hono@v3.7.6/jsx/index.ts";
-export type { HtmlEscapedString } from "https://deno.land/x/hono@v3.7.6/utils/html.ts";
-export { html } from "https://deno.land/x/hono@v3.7.6/helper/html/index.ts";
